+++ packages/assets/src/icons/IconESportsCallOfDuty.tsx
@@ -29,7 +29,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconESportsCallOfDuty =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {