+++ packages/assets/src/icons/IconNotification.tsx
@@ -32,7 +32,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconNotification =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {