+++ packages/assets/src/icons/IconSbShield.tsx
@@ -44,7 +44,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconSbShield =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {