+++ packages/assets/src/icons/IconSportsRugby.tsx
@@ -37,7 +37,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconSportsRugby =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {