+++ packages/assets/src/icons/IconUser.tsx
@@ -30,7 +30,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconUser =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {